backend/requirements.txt
blinker==1.8.2
certifi==2025.11.12
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.3
flask-cors==5.0.0
greenlet==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
PyJWT==2.10.1
MarkupSafe==2.1.5
numpy==2.0.1
pandas==2.2.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
argon2-cffi==25.1.0